+++ CHANGELOG.md
@@ -10,6 +10,7 @@ New features:
 - Detection of 45 degrees to make the diagonal runs.
 - Advise (LED indication) next advance is expected to be a diagonal.
 - Beep for long key press is longer than short key press.
+- Firmware version is shown as the first message to the serial console.
 
 ## v1.4.3 (Dec 15, 2017)